feat: extracts horizontal overflow component from base tabs
