+++ packages/palette/src/elements/HorizontalOverflow/index.ts
@@ -1,0 +1,1 @@
+export * from "./HorizontalOverflow"
