+++ packages/palette/src/elements/index.tsx
@@ -24,6 +24,7 @@ export * from "./Flex"
 export * from "./FullBleed"
 export * from "./Grid"
 export * from "./GridColumns"
+export * from "./HorizontalOverflow"
 export * from "./HTML"
 export * from "./Image"
 export * from "./Input"
